+++ lib/MIP/Recipes/Analysis/Picardtools_mergesamfiles.pm
@@ -17,13 +17,16 @@ use warnings qw{ FATAL utf8 };
 use autodie qw{ :all };
 use Readonly;
 
+## MIPs lib/
+use MIP::Constants qw{ $ASTERISK $DOT $EMPTY_STR $NEWLINE $SEMICOLON $SPACE $UNDERSCORE };
+
 BEGIN {
 
     require Exporter;
     use base qw{ Exporter };
 
     # Set the version for version checking
-    our $VERSION = 1.07;
+    our $VERSION = 1.08;
 
     # Functions and variables which can be optionally exported
     our @EXPORT_OK =
@@ -31,26 +34,18 @@ BEGIN {
 
 }
 
-## Constants
-Readonly my $ASTERISK     => q{*};
-Readonly my $DOT          => q{.};
-Readonly my $EMPTY_STRING => q{};
-Readonly my $NEWLINE      => qq{\n};
-Readonly my $SEMICOLON    => q{;};
-Readonly my $SPACE        => q{ };
-Readonly my $UNDERSCORE   => q{_};
-
 sub analysis_picardtools_mergesamfiles {
 
 ## Function : Merges all bam files using Picardtools mergesamfiles within each sampleid and files generated previously (option if provided with '-picardtools_mergesamfiles_previous_bams'). The merged files have to be sorted before attempting to merge.
 ## Returns  :
 ## Arguments: $active_parameter_href   => Active parameters for this analysis hash {REF}
-##          : $case_id               => Family id
+##          : $case_id                 => Family id
 ##          : $file_info_href          => File_info hash {REF}
 ##          : $infile_lane_prefix_href => Infile(s) without the ".ending" {REF}
 ##          : $job_id_href             => Job id hash {REF}
 ##          : $parameter_href          => Parameter hash {REF}
-##          : $recipe_name            => Program name
+##          : $profile_base_command    => Submission profile base command
+##          : $recipe_name             => Program name
 ##          : $referencefile_path      => Human genome reference file path
 ##          : $sample_id               => Sample id
 ##          : $sample_info_href        => Info on samples and case hash {REF}
@@ -71,6 +66,7 @@ sub analysis_picardtools_mergesamfiles {
 
     ## Default(s)
     my $case_id;
+    my $profile_base_command;
     my $referencefile_path;
     my $temp_directory;
     my $xargs_file_counter;
@@ -114,6 +110,11 @@ sub analysis_picardtools_mergesamfiles {
             defined     => 1,
             required    => 1,
             store       => \$parameter_href,
+            strict_type => 1,
+        },
+        profile_base_command => {
+            default     => q{sbatch},
+            store       => \$profile_base_command,
             strict_type => 1,
         },
         recipe_name => {
@@ -214,7 +215,7 @@ sub analysis_picardtools_mergesamfiles {
     my $outfile_suffix = $rec_atr{outfile_suffix};
 
     ## Extract lanes
-    my $lanes_id = join $EMPTY_STRING, @{ $file_info_href->{$sample_id}{lanes} };
+    my $lanes_id = join $EMPTY_STR, @{ $file_info_href->{$sample_id}{lanes} };
 
     ## Outpaths
     my $outsample_directory =
@@ -512,8 +513,9 @@ q{## Renaming sample instead of merge to streamline handling of filenames downst
 
         submit_recipe(
             {
-                dependency_method       => q{sample_to_sample},
+                base_command            => $profile_base_command,
                 case_id                 => $case_id,
+                dependency_method       => q{sample_to_sample},
                 infile_lane_prefix_href => $infile_lane_prefix_href,
                 job_id_chain            => $job_id_chain,
                 job_id_href             => $job_id_href,
@@ -524,7 +526,7 @@ q{## Renaming sample instead of merge to streamline handling of filenames downst
             }
         );
     }
-    return;
+    return 1;
 }
 
 sub analysis_picardtools_mergesamfiles_rio {
@@ -532,7 +534,7 @@ sub analysis_picardtools_mergesamfiles_rio {
 ## Function :  Merges all bam files using Picardtools mergesamfiles within each sampleid. The merged files have to be sorted before attempting to merge.
 ## Returns  : |$xargs_file_counter
 ## Arguments: $active_parameter_href   => Active parameters for this analysis hash {REF}
-##          : $case_id               => Family id
+##          : $case_id                 => Family id
 ##          : $FILEHANDLE              => Filehandle to write to
 ##          : $file_info_href          => The file_info hash {REF}
 ##          : $file_path               => File path
@@ -542,8 +544,8 @@ sub analysis_picardtools_mergesamfiles_rio {
 ##          : $referencefile_path      => Human genome reference file path
 ##          : $sample_id               => Sample id
 ##          : $sample_info_href        => Info on samples and case hash {REF}
-##          : $recipe_info_path       => Recipe info path
-##          : $recipe_name            => Program name
+##          : $recipe_info_path        => Recipe info path
+##          : $recipe_name             => Program name
 ##          : $temp_directory          => Temporary directory
 ##          : $xargs_file_counter      => The xargs file counter
 
@@ -711,7 +713,7 @@ sub analysis_picardtools_mergesamfiles_rio {
     my $outfile_suffix = $rec_atr{outfile_suffix};
 
     # Extract lanes
-    my $lanes_id = join $EMPTY_STRING, @{ $file_info_href->{$sample_id}{lanes} };
+    my $lanes_id = join $EMPTY_STR, @{ $file_info_href->{$sample_id}{lanes} };
 
     ## Outpaths
     my $outsample_directory =