+++ lib/MIP/Test/Fixtures.pm
@@ -20,7 +20,7 @@ use autodie qw{ :all };
 use Readonly;
 
 ## MIPs lib/
-use lib catdir( dirname($Bin), q{lib} );
+use MIP::Constants qw{ $COMMA $NEWLINE $SPACE };
 use MIP::Script::Utils qw{ help };
 
 BEGIN {
@@ -28,16 +28,11 @@ BEGIN {
     use base qw{ Exporter };
 
     # Set the version for version checking
-    our $VERSION = 1.04;
+    our $VERSION = 1.06;
 
     # Functions and variables which can be optionally exported
     our @EXPORT_OK = qw{ test_import test_log test_mip_hashes test_standard_cli };
 }
-
-## Constants
-Readonly my $COMMA   => q{,};
-Readonly my $NEWLINE => qq{\n};
-Readonly my $SPACE   => q{ };
 
 sub build_usage {
 
@@ -110,17 +105,34 @@ sub test_log {
 
 ## Function : Generate a log object and a temporary log file
 ## Returns  : $log
-## Arguments: $log_name => Name of logger
+## Arguments: $log_level => Log level
+##          : $log_name  => Name of logger
+##          : $no_screen => Don't log to screen
 
     my ($arg_href) = @_;
 
     ## Flatten argument(s)
     my $log_name;
+    my $no_screen;
+
+    ## Default(s)
+    my $log_level;
 
     my $tmpl = {
+        log_level => {
+            allow       => [qw{ DEBUG ERROR FATAL INFO TRACE WARN }],
+            default     => q{TRACE},
+            store       => \$log_level,
+            strict_type => 1,
+        },
         log_name => {
             default     => q{TEST},
             store       => \$log_name,
+            strict_type => 1,
+        },
+        no_screen => {
+            allow       => [ undef, 0, 1 ],
+            store       => \$no_screen,
             strict_type => 1,
         },
     };
@@ -129,14 +141,23 @@ sub test_log {
     use MIP::Log::MIP_log4perl qw{ initiate_logger };
 
     ## Create temp logger
-    my $test_dir      = File::Temp->newdir();
+    my $test_dir = File::Temp->newdir();
     my $test_log_path = catfile( $test_dir, q{test.log} );
+
+    ## Set default log categories
+    my @categories = ( $log_level, qw{ LogFile ScreenApp } );
+
+    ## Disable print to screen
+    if ($no_screen) {
+        @categories = ( $log_level, q{LogFile} );
+    }
 
     ## Creates log object
     my $log = initiate_logger(
         {
-            file_path => $test_log_path,
-            log_name  => $log_name,
+            categories_ref => \@categories,
+            file_path      => $test_log_path,
+            log_name       => $log_name,
         }
     );
 
@@ -161,7 +182,7 @@ sub test_mip_hashes {
     my $tmpl = {
         mip_hash_name => {
             allow => [
-                qw{ active_parameter define_parameter download_parameter file_info install_parameter io job_id pedigree recipe_parameter qc_sample_info }
+                qw{ active_parameter define_parameter dependency_tree_dna dependency_tree_rna download_parameter file_info install_parameter io job_id pedigree recipe_parameter qc_sample_info }
             ],
             defined     => 1,
             required    => 1,
@@ -188,6 +209,10 @@ sub test_mip_hashes {
         active_parameter =>
           catfile( $Bin, qw{ data test_data recipe_active_parameter.yaml } ),
         define_parameter => catfile( $Bin, qw{ data test_data define_parameters.yaml } ),
+        dependency_tree_dna =>
+          catfile( $Bin, qw{ data test_data rd_dna_initiation_map.yaml } ),
+        dependency_tree_rna =>
+          catfile( $Bin, qw{ data test_data rd_rna_initiation_map.yaml } ),
         download_parameter =>
           catfile( $Bin, qw{ data test_data download_parameters.yaml } ),
         file_info => catfile( $Bin, qw{ data test_data recipe_file_info.yaml } ),
@@ -199,6 +224,7 @@ sub test_mip_hashes {
         pedigree         => catfile( $Bin, qw{ data test_data pedigree.yaml } ),
         qc_sample_info =>
           catfile( $Bin, qw{ data test_data 643594-miptest_qc_sample_info.yaml } ),
+
     );
 
     my %hash_to_return = load_yaml(