+++ t/analysis_cnvnator.t
@@ -21,10 +21,11 @@ use Test::Trap;
 
 ## MIPs lib/
 use lib catdir( dirname($Bin), q{lib} );
+use MIP::Constants qw{ $COLON $COMMA $SPACE };
 use MIP::Test::Fixtures qw{ test_log test_mip_hashes test_standard_cli };
 
 my $VERBOSE = 1;
-our $VERSION = 1.00;
+our $VERSION = 1.01;
 
 $VERBOSE = test_standard_cli(
     {
@@ -35,9 +36,6 @@ $VERBOSE = test_standard_cli(
 
 ## Constants
 Readonly my $CNV_BIN_SIZE => 1000;
-Readonly my $COLON        => q{:};
-Readonly my $COMMA        => q{,};
-Readonly my $SPACE        => q{ };
 
 BEGIN {
 
@@ -64,10 +62,10 @@ diag(   q{Test analysis_cnvnator from Cnvnator.pm v}
       . $SPACE
       . $EXECUTABLE_NAME );
 
-my $log = test_log( { log_name => q{MIP}, } );
+my $log = test_log( { log_name => q{MIP}, no_screen => 1, } );
 
 ## Given analysis parameters
-my $recipe_name    = q{cnvnator};
+my $recipe_name = q{cnvnator};
 my $slurm_mock_cmd = catfile( $Bin, qw{ data modules slurm-mock.pl } );
 
 my %active_parameter = test_mip_hashes(